dont launch lambda server in oss
